Smart Thief Up Movement Implemented
Smart Thief can get to the nearest interactable item placed above it with valid moves.
If the item is placed on an unreachable tile, Smart Thief will get as close as possible and stop.

Changed movementTimerPass in all classes that use it into integers since they represent milliseconds

+++ src/main/java/com/example/cs230assignment/FlyingAssassin.java
@@ -2,7 +2,7 @@ package com.example.cs230assignment;
 
 public class FlyingAssassin extends NPC {
 
-    public FlyingAssassin(double movementTimerPass, int x, int y, int xdir,
+    public FlyingAssassin(int movementTimerPass, int x, int y, int xdir,
             int ydir) {
         super("Flying Assassin", movementTimerPass, x, y);
         super.coordChange = new int[] {xdir,ydir};
