+++ src/main/java/com/example/cs230assignment/NPC.java
@@ -1,13 +1,13 @@
 package com.example.cs230assignment;
 
 public class NPC extends Character {
-    private double movementTimer;
+    private int movementTimer;
     protected int[] coordChange;
     protected Board currentBoard;
 
     // thinking about adding move into here aswell as character for the basic
     // movement of the NPCS
-    public NPC(String name, double movementTimerPass, int x, int y) {
+    public NPC(String name, int movementTimerPass, int x, int y) {
         super(name, x, y);
         this.movementTimer = movementTimerPass;
     }
